De world bevat een tijd.
git-svn-id: file:///home/kommusoft/tmpgit/swopsvn@349 c49496ef-df08-40b3-bdd7-5dbf4c3ba35c

+++ ProjectSWOP20102011/src/projectswop20102011/domain/lists/TimeSensitiveList.java
@@ -65,7 +65,7 @@ public class TimeSensitiveList implements Iterable<TimeSensitive> {
 	 * @throws InvalidDurationException
 	 *		If the given duration is invalid. I.e. the duration is negative.
 	 */
-	public void timeAhead(int seconds) throws InvalidDurationException {
+	public void timeAhead(long seconds) throws InvalidDurationException {
 		for (TimeSensitive ts : this) {
 			ts.timeAhead(seconds);
 		}
